refactor,tests: section deltas to package
