+++ tests/calc/sections/deltas/conftest.py
@@ -1,0 +1,11 @@
+"""Test configuration for tests/calc/sections/deltas."""
+
+__authors__ = (
+    'Michele Riva (@michele-riva)',
+    )
+__copyright__ = 'Copyright (c) 2019-2025 ViPErLEED developers'
+__created__ = '2025-07-04'
+__license__ = 'GPLv3+'
+
+
+_MODULE = 'viperleed.calc.sections.deltas'
